Make GL stuff run with lower GLSL version number
	modified:   ../cvt/gl/GLProgram.cpp
	modified:   ../cvt/gl/GLProgram.h
	modified:   ../cvt/gl/shader/basic.frag
	modified:   ../cvt/gl/shader/text.frag
	modified:   ../cvt/gui/GFXGL.cpp

git-svn-id: https://svnknoll.informatik.tu-muenchen.de/kloses/visionTools@406 f7bcf5df-4c73-0410-a0d8-b61c104701fe

+++ cvt/gl/GLProgram.h
@@ -12,8 +12,8 @@ namespace cvt {
 			~GLProgram();
 			void build( const char* vertsrc, const char* fragsrc );
 			void bind();
-			void bindAttribLocation( const char *name, size_t location );
-			size_t getAttribLocation( const char* name );
+			void bindAttribLocation( const char *name, unsigned int location );
+			unsigned int getAttribLocation( const char* name );
 			void setArg( size_t n, IFilterParameter* p );
 			void setArg( const char* name, IFilterParameter* p );
 			void setArg( const char* name, int i );
